refactor SGMCMC posterior into a top-level class
